+++ lib/pdfmd.rb
@@ -36,9 +36,7 @@
 # * Subject
 # * Keywords (optional)
 #
-# TODO: Include password protected PDF documents as well
 # TODO: Fix broken PDF files automatically
-# TODO: Enable logging in 'edit'
 # TODO: Read this: http://lostechies.com/derickbailey/2011/04/29/writing-a-thor-application/
 # TODO: ... and this: http://blog.paracode.com/2012/05/17/building-your-tools-with-thor/
 # gs \
@@ -59,7 +57,7 @@ require "i18n"
 require 'pathname'
 require 'logger'
 
-VERSION = '1.7.0'
+VERSION = '1.8.0'
 
 # Include general usage methods
 require_relative('pdfmd/methods.rb')