+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Version 1.8.0
+- Added Support for password protected pdf files in command 'show' and 'edit'
+- Cleaned up renaming key-string and added all string for NO,EN an DE language.
+- Cleaned the output of `pdfmd config`. 
+- Removed some TODOs
+- Bugfix in the rename command
+- Updated Tests
+- Removed special characters in subject when used in the filename (bug)
+
 # Version 1.7.0
 - Rename option in command 'edit' on the Shell now overwrites the hiera setting.
 - The command 'show' supports multiple output formats.